# perf(Designer): Moved `canDrop` calculations to a conditional component