## Commit Type
<!-- Select one -->
- [x] perf - Performance improvement

## Risk Level
<!-- Select one based on potential impact -->
- [x] Low - Minor changes, limited scope
- [ ] Medium - Moderate changes, some user impact
- [ ] High - Major changes, significant user/system impact

## What & Why
<!-- Brief context: What does this change and why? -->
Moved logic determining the `canDrop` flag on edge drop targets to a conditional component.
Logic now only runs when we need to determine the flag instead of on-render, greatly reducing rending time.
Starting a drag is now slightly slower, but when testing a very zoomed out flow with ~300 nodes on screen it was a barely noticeable difference.
Due to the infrequency of this interaction compared to every time an edge is rendered, I think this is an acceptable tradeoff.

The following tests were run on a 500 node test flow to highlight performance differences.

| [3 test averages]   | Before | After | % Change |
|---------------------|--------|-------|----------|
| Expanding 500 nodes | 8.3s   | 6.3s  | 34% faster   |
| Single edge rendering time | 12.9ms   | 6.6s  | 95% faster   |

## Impact of Change
<!-- Who/what is affected? -->
- **Users**: Navigating workflows will feel faster, initial loads will be faster as well
- **Developers**: No change
- **System**: Improved perf when rendering edges

## Test Plan
<!-- How was this tested? -->
- [ ] Unit tests added/updated
- [ ] E2E tests added/updated
- [x] Manual testing completed
- [ ] Tested in: <!-- environments/scenarios -->

## Contributors
<!-- Tag team members who contributed ideas, reviews, or implementation -->
@rllyy97

## Screenshots/Videos
<!-- Visual changes only -->